requirements.txt
absl-py==1.0.0
numpy==1.21.4
tensorflow==2.7.0
opencv-python==4.5.4.60
requests==2.26.0
selenium==4.1.0
pycocotools==2.0.3